test: Add additional interference_mutex unit test
The existing unit tests for interference_mutex will pass bad
implementations because they only test actions in one direction.
That is, they check if any preconditions from the first argument
are mutex with any effects of the second, but there is no test
for the preconditions of the second against the effects of the
first.