test: Add additional inconsistent_support_mutex unit test
The existing unit tests for inconsistent_support_mutex will pass
bad implementations because they do not test for cases where
a single parent action is capable of producing all the states.
In such cases, it does not matter if the parent actions are mutex,
the states do not satisfy the inconsistent support mutex condition.